prov/rxm: Updates to rxm provider to suppress warnings

+++ include/fi.h
@@ -163,7 +163,7 @@ static inline size_t fi_get_aligned_sz(size_t size, size_t alignment)
 
 uint64_t fi_tag_bits(uint64_t mem_tag_format);
 uint64_t fi_tag_format(uint64_t tag_bits);
-int fi_size_bits(uint64_t num);
+uint8_t fi_size_bits(uint64_t num);
 
 int ofi_send_allowed(uint64_t caps);
 int ofi_recv_allowed(uint64_t caps);
